# docs/swagger.yaml
basePath: /
definitions:
...
  resource.InputUser:
    properties:
      age:
        example: 20
        minimum: 9
        type: integer
      email:
        example: testing@gmail.com
        type: string
      password:
        example: test123
        minLength: 6
        type: string
      username:
        example: testing
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  resource.LoginUser:
    properties:
      email:
        example: testing@gmail.com
        type: string
      password:
        example: test123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  resource.UpdateUser:
    properties:
      email:
        example: testing@gmail.com
        type: string
      username:
        example: testing
        type: string
    required:
    - email
    - username
    type: object
info:
  contact:
    email: kadekbintanga@gmail.com
    name: API Support
  description: This is a service for final project
  license:
    name: Apache 2.0
  termsOfService: http://swagger.io/terms/
  title: Final Project by Bintang
  version: "1.0"
paths:
...
  /api/v1/user:
    delete:
      consumes:
      - application/json
      description: delete a user with the token
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update a user with the input payload
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.UpdateUser'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a user
      tags:
      - User
...
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      description: Login User  with the input payload
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: Login a user
      tags:
      - User
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      description: Create a new user with the input payload
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.InputUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: Create a new user
      tags:
      - User
swagger: "2.0"